Refactor systemfields
